backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
alembic==1.15.2
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
redis==5.2.1
celery==5.4.0
python-jose==3.3.0
cryptography==44.0.2
python-multipart==0.0.20
